Fix unit test for block of 0.1a version
- Solve recursive import problem
- Add utils module
