+++ tests/test_loopchain_block.py
@@ -21,7 +21,8 @@ from icondbtools.loopchain_block import LoopchainBlock
 
 
 class TestLoopchainBlock(unittest.TestCase):
-    def test_from_dict(self):
+
+    def test_block_v0_1a_from_dict(self):
         version = '0.1a'
         prev_block_hash = bytes.fromhex(
             'f578312788010043e7a0b959e0c0dbca87c4e7a90331ac59a91fb2fdb62c8640')
@@ -33,41 +34,33 @@ class TestLoopchainBlock(unittest.TestCase):
         height = 59669
         peer_id = Address.from_string('hx667e748e93a5a4e1d61c92c482577888e8c35c9d')
         commit_state = bytes.fromhex('4b57f72f29ebdf5d5543e12ea63e684ab6636c53ca24600f0208e410f0311447')
+        confirmed_transaction_list = [{
+            'from': 'hxeb56a51667eb0491bd1308426865193a9684908a',
+            'to': 'hx5c328b010e4ef0f81670ef48eb1b903aac1443e2',
+            'value': '0x8ac7230489e80000',
+            'version': '0x3',
+            'nid': '0x1',
+            'stepLimit': '0x186a0',
+            'timestamp': '0x57648a309f940',
+            'signature': 'KglKpRzx1kRCmfOEnU7ltIa2PFaJDhHElowNFxDeXgNP7ksttPCT3fnSR4PpoPpkVVh3wprQVZr+KCPK6r7uRwA=',
+            'txHash': '30cdeb912b4acdac5c2cdd885983648c366ec0d99ce99a841bb4620f34f9b5c9'
+        }]
 
         block_dict = {
-			'version': version,
-			'prev_block_hash': prev_block_hash.hex(),
-			'merkle_tree_root_hash': merkle_tree_root_hash.hex(),
-			'time_stamp': timestamp,
-			'confirmed_transaction_list': [{
-				'from': 'hxeb56a51667eb0491bd1308426865193a9684908a',
-				'to': 'hx5c328b010e4ef0f81670ef48eb1b903aac1443e2',
-				'value': '0x8ac7230489e80000',
-				'version': '0x3',
-				'nid': '0x1',
-				'stepLimit': '0x186a0',
-				'timestamp': '0x57648a309f940',
-				'signature': 'KglKpRzx1kRCmfOEnU7ltIa2PFaJDhHElowNFxDeXgNP7ksttPCT3fnSR4PpoPpkVVh3wprQVZr+KCPK6r7uRwA=',
-				'txHash': '30cdeb912b4acdac5c2cdd885983648c366ec0d99ce99a841bb4620f34f9b5c9'
-			}],
-			'block_hash': block_hash.hex(),
-			'height': height,
-			'peer_id': str(peer_id),
-			'signature': 'M97tADsivw0qtzD0qJUDjIc/ki6Zf5HatoAnOTgpi2dNmP5WvJccAVpD86mjxxNOtdnFnO009iwnvw6yU0RBNAA=',
-			'commit_state': {'icon_dex': commit_state.hex()}
-		}
+            'version': version,
+            'prev_block_hash': prev_block_hash.hex(),
+            'merkle_tree_root_hash': merkle_tree_root_hash.hex(),
+            'time_stamp': timestamp,
+            'confirmed_transaction_list': confirmed_transaction_list,
+            'block_hash': block_hash.hex(),
+            'height': height,
+            'peer_id': str(peer_id),
+            'signature': 'M97tADsivw0qtzD0qJUDjIc/ki6Zf5HatoAnOTgpi2dNmP5WvJccAVpD86mjxxNOtdnFnO009iwnvw6yU0RBNAA=',
+            'commit_state': {'icon_dex': commit_state.hex()}
+        }
 
         loopchain_block = LoopchainBlock.from_dict(block_dict)
-        self.assertEqual(version, loopchain_block.version)
-        self.assertEqual(prev_block_hash, loopchain_block.prev_block_hash)
-        self.assertEqual(merkle_tree_root_hash, loopchain_block.merkle_tree_root_hash)
+        self.assertEqual(prev_block_hash, loopchain_block.prev_hash)
         self.assertEqual(timestamp, loopchain_block.timestamp)
-        self.assertEqual(block_hash, loopchain_block.block_hash)
+        self.assertEqual(block_hash, loopchain_block.hash)
         self.assertEqual(height, loopchain_block.height)
-        self.assertEqual(peer_id, loopchain_block.peer_id)
-        self.assertEqual(commit_state, loopchain_block.commit_state)
-
-        block_dict['peer_id'] = ''
-        loopchain_block = LoopchainBlock.from_dict(block_dict)
-        self.assertIsNone(loopchain_block.peer_id)
-
